Add hot reload endpoint
This commit brings a new endpoint which allows designated users to hot reload the configuration file for the application. This also brings along the concept of a disconnected client, where a queue becomes inaccessible to an existing client.

This commit also brings a simple surface level configuration validator. It does not validate types, however it does ensure that only valid keys are provided which should catch most real world issues.